Use size_t for decode API input size parameters.
Use a type more appropriate for buffer sizes that more
cleanly integrates with I/O APIs. Removes some casting
related to size_t use with C file I/O APIs (fread mainly).

This changes the call signatures of aom_codec_decode() and
aom_codec_peek_stream_info(), and updates the internal
implementations. The single change to all API entries is
and internal hooks is the change from "unsigned int" to
size_t for the data_sz parameter.

+++ aom/aom_decoder.h
@@ -162,8 +162,7 @@ aom_codec_err_t aom_codec_dec_init_ver(aom_codec_ctx_t *ctx,
  *     buffer was too short.
  */
 aom_codec_err_t aom_codec_peek_stream_info(aom_codec_iface_t *iface,
-                                           const uint8_t *data,
-                                           unsigned int data_sz,
+                                           const uint8_t *data, size_t data_sz,
                                            aom_codec_stream_info_t *si);
 
 /*!\brief Return information about the current stream.
@@ -212,7 +211,7 @@ aom_codec_err_t aom_codec_get_stream_info(aom_codec_ctx_t *ctx,
  *         for recoverability capabilities.
  */
 aom_codec_err_t aom_codec_decode(aom_codec_ctx_t *ctx, const uint8_t *data,
-                                 unsigned int data_sz, void *user_priv);
+                                 size_t data_sz, void *user_priv);
 
 /*!\brief Decoded frames iterator
  *
